Fix signed/unsigned comparison warnings (C4388/C4389)
PPgSecurity: use size_t for the _countof loop index. TransferWnd: cast
the guaranteed-non-negative iCategory to UINT for the comparison against
the unsigned m_curTab. Both are benign mismatches; the fixes make the
signedness explicit without behavior change.

+++ srchybrid/PPgSecurity.cpp
@@ -288,7 +288,7 @@ void CPPgSecurity::SeedDefaultIPFilterUpdateUrls()
 	if (m_pacIPFilterURL == NULL || !m_pacIPFilterURL->IsBound() || m_pacIPFilterURL->GetItemCount() > 0)
 		return;
 
-	for (int i = 0; i < _countof(s_apszDefaultIPFilterUpdateUrls); ++i)
+	for (size_t i = 0; i < _countof(s_apszDefaultIPFilterUpdateUrls); ++i)
 		m_pacIPFilterURL->AddItem(s_apszDefaultIPFilterUpdateUrls[i], -1);
 }
